requirements.txt
docx2txt==0.8
numpy==1.21.6
numpy==1.23.5
numpy==1.25.2
opencv_python==4.8.0.74
Pillow==9.5.0
Pillow==10.0.1
PyPDF2==3.0.1
pyperclip==1.8.2
streamlit==1.26.0
